chore: move to_newick implementation to newick.rs
